+++ src/pages/item/ItemPage.js
@@ -1,8 +1,8 @@
 import React from "react";
-import BestProducts from "../component/BestProducts";
-import Products from "../component/Products";
-import "../css/itemPage.css";
-import useSetNumOfItemsToShow from "../hooks/setNumOfItemsToShow";
+import BestProducts from "./BestProducts";
+import Products from "./Products";
+import "./itemPage.css";
+import useSetNumOfItemsToShow from "../../hooks/setNumOfItemsToShow";
 
 const ItemPage = () => {
   const [bestProductsNumOfItemsToShow, ProductsNumOfItemsToShow] =
