+++ src/pages/item/itemPage.css
@@ -1,3 +1,4 @@
+
 .itemPage-main{
     display: grid;
     grid-template: 2.4rem 123.4rem 11.3rem / 1fr 120rem 1fr;
