+++ src/pages/item/products.css
@@ -22,7 +22,7 @@
   border-radius: 1.2rem;
   background-color: var(--coolGray100);
   border: none;
-  background-image: url("../image/search_icon.png");
+  background-image: url("../../image/search_icon.png");
   background-position: 1.6rem 45%;
   background-repeat: no-repeat;
   padding-left: 4rem;
@@ -54,7 +54,7 @@
   width: 13rem;
   padding-left: 2rem;
   text-align: left;
-  background-image: url("../image/ic_arrow_down.svg");
+  background-image: url("../../image/ic_arrow_down.svg");
   background-repeat: no-repeat;
   background-position: 8.6rem;
   font-size: 1.6rem;
@@ -142,7 +142,7 @@
     padding: 0;
     font-size: 0;
     width: 4.2rem;
-    background-image: url("../image/ic_sort.svg"); /* 이미지 경로는 실제 이미지 파일의 경로로 대체해주세요 */
+    background-image: url("../../image/ic_sort.svg"); /* 이미지 경로는 실제 이미지 파일의 경로로 대체해주세요 */
     background-size: contain; /* 배경 이미지를 요소에 맞게 조절합니다 */
     background-repeat: no-repeat; /* 배경 이미지 반복을 막습니다 */
     background-position: left center;
